requirements.txt
langgraph
langchain
langchain_openai
langchain_community
langchain_core
python-dotenv
langchain_pinecone
pinecone
jq
langchain-google-genai
pandas
langchain_anthropic
langchain_groq
tavily-python
ipython
fastapi[standard]